requirements.txt
fastapi==0.62.0
gunicorn==20.0.4
uvicorn==0.13.1
uvloop==0.14.0
httptools==0.1.1
python-dateutil==2.8.1
